Remove old unlock command
